pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wiki-vault"
version = "0.1.0"
description = "A CLI tool that turns raw sources into structured, interlinked Obsidian wikis"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "python-frontmatter>=1.0",
    "pyyaml>=6.0",
    "requests>=2.31",
    "readability-lxml>=0.8",
    "lxml>=5.0",
    "markdownify>=0.13",
]

[project.scripts]
wiki-vault = "wiki_vault.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
